requirements.txt
async-timeout==4.0.2
Authlib==0.15.6
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
dataclasses-json==0.5.7
Flask==2.2.2
Flask-Login==0.6.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==0.21.1
python-twitter-v2==0.8.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
tomli==2.0.1
tweepy==4.12.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
